+++ derive/CHANGELOG.md
@@ -1,5 +1,5 @@
-Unreleased
-----------
+0.3.0
+-----
 - Added support for `gui(Event = ...)` attribute
 - Dropped "Gui" prefix from custom derive macros
 - Downgraded dependency to `gui` to a dev-dependency